add config.yaml
