+++ backend/src/config.yaml
@@ -1,17 +1,18 @@
+
 # Hugging Face: https://huggingface.co/sentence-transformers/allenai-specter
-embedding_model = "sentence-transformers/allenai-specter"
+embedding_model : "sentence-transformers/allenai-specter"
 
 # dataset link
-papers_data_url = "http://www.lri.fr/owncloud/index.php/s/OO987IvsoKwWI3l/download"
+papers_data_url : "http://www.lri.fr/owncloud/index.php/s/OO987IvsoKwWI3l/download"
 
 # Number of search results to show.
-top_k = 5
+top_k : 5
 
 # Protein file path relative to root.
-papers_data_path = "data/papers.csv"
+papers_data_path : "data/papers.csv"
 
 # Prints logs to command line if true.
-print_logs = True
+print_logs : True
 
 # Search service port
-search_port = 8020
+search_port : 8020
