+++ backend/src/flows.py
@@ -1,7 +1,7 @@
 from jina import Flow
 
+from config import config
 from executors import SpecterExecutor
-from backend_config import top_k, search_port
 
 # Using a standard indexer: https://hub.jina.ai/executor/zb38xlt4
 indexer = "jinahub://SimpleIndexer"
@@ -21,14 +21,14 @@ def index_flow():
 
 def search_flow():
     flow = (
-        Flow(port_expose=search_port, protocol="http")
+        Flow(port_expose=config["search_port"].get(), protocol="http")
         .add(uses=SpecterExecutor)
         .add(
             uses=indexer,
             uses_with={
                 "match_args": {
                     "metric": "cosine",
-                    "limit": top_k,
+                    "limit": config["top_k"].get(),
                 },
             },
         )
